requirements.txt
streamlit>=1.31
psycopg2-binary>=2.9
python-dotenv>=1.0
requests>=2.31
sqlparse>=0.4.4
